feat: add a CGI script for viewing data
Replace the virtual environment for Python 3.13 with a
virtual environment for Python 3.10, the version available on
the shared hosting server. Modify the database connection
code to eliminate Python 3.13-specific code. Modify .gitignore
to ignore the new virtual environment in venv310.

Create view.cgi for viewing the data in the database.

Add sample-form.html as an example of modern HTML5
methods for styling a form.

Update README.md with additional information about the project.